Add note about gh-pages content-type workaround
Github Pages serves .md files without charset=utf-8 in the content
type, so some browsers will consider such files to be encoded as
ISO-8859-1. Of course, for non-ascii utf-8 encoded files, this will
make a mess.

The workaround is to download the .md file and open that file
instead. The problem is specifically for downloading the raw .md
file as published on Github Pages. If you view the file in Github's
repository browser, the markdown rendered as html will probably
look fine.