Add footer link for license
This makes the footer's mention of the license into a clickable
link pointing to the plaintext license in this repo.

To convince github pages to use the text/plain content type when
serving a response for the license link, I had to rename LICENSE
to LICENSE.txt. Otherwise Chrome would download it instead of
opening it in a tab as a text file.